requirements.txt
attrs==25.1.0
boto3==1.36.14
botocore==1.36.14
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
jmespath==1.0.1
numpy==2.2.2
outcome==1.3.0.post0
pandas==2.2.3
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
s3transfer==0.11.2
selenium==4.28.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.37
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
